langgraphapp/requirements.txt
langchain
langgraph
python-dotenv
langchain-community
dashscope